backend/requirements.txt
# web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
pgvector==0.3.4

# auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# background tasks
celery[redis]==5.4.0
redis==5.1.1

# ai and document processing
openai==1.51.0
tiktoken==0.7.0
pypdf2==3.0.1
python-docx==1.1.2

# storage
boto3==1.35.28

# utilities
structlog==24.4.0
tenacity==9.0.0
httpx==0.27.2

# testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
factory-boy==3.3.1
ruff==0.6.9
